Ya casi

+++ api/Hospedaje/database/migrations/2024_02_10_063330_updated_comments_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('comments', function (Blueprint $table) {
             $table->id();
             $table->string('messaje');
-            $table->unsignedBigInteger('user_id');
+            $table->int('user_id');
             $table->foreign('user_id')->references('id')->on('users');
             $table->timestamps();
         });
